+++ symmetric.py
@@ -1,24 +1,32 @@
-from cryptography.hazmat.primitives import padding, hashes
-from cryptography.hazmat.primitives.asymmetric import padding as aspadding
+import os
+from cryptography.hazmat.primitives import padding
 from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.primitives.serialization import load_pem_private_key
 import logging
 
 
-def load_private_key(private_pem: str):
+logger = logging.getLogger()
+logger.setLevel('INFO')
+
+
+def symmetric_encrypt(key: bytes, text: bytes, len: int) -> bytes:
     """
-    Функция десериализации закрытого ключа
-    :param private_pem: название файла
-    :return: закрытый ключ
+    Функция шифрования текста алгоритмом симметричного шифрования Camellia
+    :param len: длина ключа
+    :param text: шифруемый текст
+    :param key: ключ
+    :return: зашифрованный текст
     """
     try:
-        with open(private_pem, 'rb') as pem_in:
-            private_bytes = pem_in.read()
-        private_key = load_pem_private_key(private_bytes, password=None)
-        logging.info(f'Закрытый ключ десериализован из {private_pem}')
+        padder = padding.ANSIX923(len).padder()
+        padded_text = padder.update(text) + padder.finalize()
+        iv = os.urandom(16)
+        cipher = Cipher(algorithms.Camellia(key), modes.CBC(iv))
+        encryptor = cipher.encryptor()
+        encrypted_text = encryptor.update(padded_text) + encryptor.finalize()
+        logging.info(f'Текст зашифрован алгоритмом Camellia')
     except OSError as err:
-        logging.warning(f'{err} Ошибка при десериализации закрытого ключа из {private_pem}')
-    return private_key
+        logging.warning(f'{err} Ошибка при симметричном шифровании!')
+    return iv + encrypted_text
 
 
 def load_symmetric_key(file_name: str) -> bytes:
@@ -55,20 +63,3 @@ def symmetric_decrypt(key: bytes, cipher_text: bytes, len: int) -> bytes:
     except OSError as err:
         logging.warning(f'{err} Ошибка при симметричном дешифровании!')
     return unpadded_text
-
-
-def asymmetric_decrypt(private_key, cipher_text: bytes) -> bytes:
-    """
-    Функция расшифровки асимметрично зашифрованного текста при помощи закрытого ключа
-    :param text: зашифрованный текст
-    :param private_key: закрытый ключ
-    :return: расшифрованный текст
-    """
-    try:
-        decrypted_text = private_key.decrypt(cipher_text,
-                                             aspadding.OAEP(mgf=aspadding.MGF1(algorithm=hashes.SHA256()),
-                                                          algorithm=hashes.SHA256(), label=None))
-        logging.info(f'Асимметрично зашифрованный текст расшифрован')
-    except OSError as err:
-        logging.warning(f'{err} Ошибка при асимметричном дешифровании!')
-    return decrypted_text
